fixed id to i64

+++ src/app/auth/models.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Debug, Clone, Deserialize)]
 pub struct AuthUser {
-    pub user_id: i32,
+    pub user_id: i64,
     pub username: String,
-    pub company_id: i32,
+    pub company_id: i64,
 }
